+++ blockchain/dkg/dkgtasks/dispute_missing_key_shares_task.go
@@ -159,7 +159,7 @@ func (t *DisputeMissingKeySharesTask) getAccusableParticipants(ctx context.Conte
 	for _, p := range t.State.Participants {
 		_, isValidator := validatorsMap[p.Address]
 		if isValidator && (p.Nonce != t.State.Nonce ||
-			p.Phase != uint8(objects.KeyShareSubmission) ||
+			p.Phase != objects.KeyShareSubmission ||
 			(p.KeyShareG1s[0].Cmp(big.NewInt(0)) == 0 &&
 				p.KeyShareG1s[1].Cmp(big.NewInt(0)) == 0) ||
 			(p.KeyShareG1CorrectnessProofs[0].Cmp(big.NewInt(0)) == 0 &&
